bug reporting improvements
